Move search route to own file
